rag-chatbot/requirements.txt
fastapi[all]
uvicorn
langchain
langchain-community
langchain-openai
langchain-groq
fastembed
qdrant_client
python-dotenv
chainlit==1.3.2